+++ src/components/pages/main/PitchTuner.svelte
@@ -1,20 +1,26 @@
 <script lang="ts">
   import { PitchDetector } from 'pitchy';
-  import { NOTES, type note_types } from './constants';
+  import {
+    NOTES,
+    type note_types,
+    getNoteNumberFromPitch,
+    getScaleFromNoteNumber,
+    getDetuneFromPitch
+  } from './constants';
   import { onMount, type Snippet } from 'svelte';
   import { FiPlay, FiRefreshCcw } from 'svelte-icons-pack/fi';
   import Icon from '~/tools/Icon.svelte';
   import { BiStopCircle } from 'svelte-icons-pack/bi';
-  import { fade, scale, slide } from 'svelte/transition';
-  import { BsChevronDown, BsChevronUp, BsMic } from 'svelte-icons-pack/bs';
+  import { fade, slide } from 'svelte/transition';
+  import { BsMic } from 'svelte-icons-pack/bs';
   import { delay } from '~/tools/delay';
   import { cl_join } from '~/tools/cl_join';
-  import PitchDisplay from './PitchDisplay.svelte';
-  import { Popover } from '@skeletonlabs/skeleton-svelte';
   import { indactivity_timeout } from './inactivity';
   import ms from 'ms';
   import { Microphone } from '@mozartec/capacitor-microphone';
   import { Tabs } from '@skeletonlabs/skeleton-svelte';
+  import CircularScale from './circular/CircularScale.svelte';
+  import PitchTimeGraph from './time_graph/PitchTimeGraph.svelte';
 
   let {
     selected_device = $bindable(),
@@ -40,16 +46,23 @@
   let update_interval: NodeJS.Timeout | null = null;
   let mic_stream: MediaStream | null = null;
 
-  let orientation_popup_status = $state(false);
-  let Sa_at_popup_status = $state(false);
-
-  const AUDIO_INFO_UPDATE_INTERVAL = 100;
+  const AUDIO_INFO_UPDATE_INTERVAL = 50;
   const GRAPH_TOTAL_TIME_MS = 8000;
 
   const FFT_SIZE = Math.pow(2, 12); // 4096
 
   // Pitch history for time graph
   const MAX_PITCH_HISTORY_POINTS = Math.floor(GRAPH_TOTAL_TIME_MS / AUDIO_INFO_UPDATE_INTERVAL); // 25 points
+
+  let started = $state(false);
+  let audio_info = $state<{
+    pitch: number;
+    clarity: number;
+    note: string;
+    scale: number;
+    detune: number;
+  } | null>(null);
+
   let pitch_history = $state<
     Array<{
       time: number;
@@ -58,9 +71,6 @@
       clarity: number;
     }>
   >([]);
-
-  // Note order with A at top (instead of C)
-  const NOTE_ORDER_A_FIRST = ['A', 'A#', 'B', 'C', 'C#', 'D', 'D#', 'E', 'F', 'F#', 'G', 'G#'];
 
   const get_audio_devices = async (show_loading = true) => {
     if (show_loading) device_list_loaded = false;
@@ -76,38 +86,6 @@
     get_audio_devices();
     indactivity_timeout(ms('30mins'), Stop);
   });
-
-  let started = $state(false);
-
-  let audio_info = $state<{
-    pitch: number;
-    clarity: number;
-    note: string;
-    scale: number;
-    detune: number;
-  } | null>(null);
-
-  const getNoteNumberFromPitch = (frequency: number) => {
-    const noteNum = 12 * (Math.log(frequency / 440) / Math.log(2));
-    return Math.round(noteNum) + 69;
-  };
-
-  const getNoteFrequency = (note: number) => {
-    return 440 * Math.pow(2, (note - 69) / 12);
-  };
-
-  // Convert frequency to y-position for graph based on semitone offset within one octave
-  const frequencyToYPosition = (frequency: number) => {
-    const A0_freq = 27.5;
-    // Compute continuous semitone offset from A0
-    const semitoneOffset = 12 * (Math.log(frequency / A0_freq) / Math.log(2));
-    // Fractional semitone within one octave
-    const semitoneInOctave = ((semitoneOffset % 12) + 12) % 12;
-    // Normalize so semitone 0 (A) maps just above axis, and semitone 11 (G#) maps to top
-    const normalized = ((semitoneInOctave + 1) / 12) * 100;
-    // Clamp to [0,100]
-    return Math.min(Math.max(normalized, 0), 100);
-  };
 
   const Stop = () => {
     started = false;
@@ -200,22 +178,22 @@
       const input = new Float32Array(detector.inputLength);
 
       clearInterval(update_interval!);
-      pitch_history = []; // Reset pitch history when starting
+      pitch_history = [];
 
-      // update every 100ms
-      update_interval = setInterval(() => {
+      function updateAudioInfo() {
         analyzer_node?.getFloatTimeDomainData(input);
         const [pitch, clarity] = detector.findPitch(input, audio_context!.sampleRate);
         // console.log(pitch, clarity);
-        const rawNote = getNoteNumberFromPitch(pitch);
-        const noteName = NOTES[rawNote % 12];
+        // console.log(pitch, clarity);
+        const rawNoteNumber = getNoteNumberFromPitch(pitch);
+        const noteName = NOTES[rawNoteNumber % 12];
         // console.log(pitch);
         const currentAudioInfo = {
           pitch: Math.round(pitch * 10) / 10,
           clarity: Math.round(clarity * 100),
           note: noteName,
-          scale: Math.floor(rawNote / 12) - 1,
-          detune: Math.floor((1200 * Math.log(pitch / getNoteFrequency(rawNote))) / Math.log(2))
+          scale: getScaleFromNoteNumber(rawNoteNumber),
+          detune: getDetuneFromPitch(pitch, rawNoteNumber)
         };
 
         audio_info = currentAudioInfo;
@@ -233,7 +211,9 @@
         if (pitch_history.length > MAX_PITCH_HISTORY_POINTS) {
           pitch_history = pitch_history.slice(-MAX_PITCH_HISTORY_POINTS);
         }
-      }, AUDIO_INFO_UPDATE_INTERVAL);
+      }
+      updateAudioInfo();
+      update_interval = setInterval(updateAudioInfo, AUDIO_INFO_UPDATE_INTERVAL);
       started = true;
     } catch (error) {
       console.log('error in Start-->', error);
@@ -247,17 +227,6 @@
       Start(); // Restart with new device if already running
     }
   };
-
-  // Prepare data for LayerChart time graph
-  const graphData = $derived(
-    pitch_history.map((point, index) => ({
-      x: index * AUDIO_INFO_UPDATE_INTERVAL, // Time in milliseconds from start
-      y: frequencyToYPosition(point.pitch), // Y position (0-100)
-      pitch: point.pitch,
-      note: point.note,
-      clarity: point.clarity
-    }))
-  );
 </script>
 
 {#if !started}
@@ -303,7 +272,6 @@
 {/snippet}
 
 {#if audio_info && started}
-  {@const { clarity, pitch } = audio_info}
   <Tabs
     value={selected_pitch_display_type}
     onValueChange={(e) =>
@@ -316,246 +284,22 @@
     {/snippet}
     {#snippet content()}
       <Tabs.Panel value="circular_scale">
-        <div class="mb-4 select-none">
-          <div class="flex flex-col items-center justify-center space-y-2 sm:space-y-3">
-            <div class="mt-2 outline-hidden select-none sm:mt-4">
-              <div class="flex items-start justify-center space-x-4">
-                <Popover
-                  open={Sa_at_popup_status}
-                  onOpenChange={(e) => (Sa_at_popup_status = e.open)}
-                  contentBase="card z-50 space-y-2 p-2 rounded-lg shadow-xl dark:bg-surface-900 bg-slate-100"
-                >
-                  {#snippet trigger()}
-                    <div class="text-center outline-hidden">
-                      <span class="mr-2 font-bold"><span>S</span> at</span>
-                      {#if !Sa_at_popup_status}
-                        <Icon src={BsChevronDown} class="text-base" />
-                      {:else}
-                        <Icon src={BsChevronUp} class="text-base" />
-                      {/if}
-                      {selected_Sa_at}
-                    </div>
-                  {/snippet}
-                  {#snippet content()}
-                    <div class="grid grid-cols-4 gap-x-2 gap-y-1 sm:grid-cols-6 sm:gap-x-2">
-                      {#each NOTES as _, i}
-                        {@const note = NOTES[(i + 9) % NOTES.length]}
-                        <button
-                          class={cl_join(
-                            'gap-0 rounded-md px-1 py-1 text-sm font-semibold text-white sm:text-base',
-                            selected_Sa_at === note
-                              ? 'bg-primary-500 dark:bg-primary-600'
-                              : 'bg-slate-400 hover:bg-primary-500/80 dark:bg-slate-800 dark:hover:bg-primary-600/80'
-                          )}
-                          onclick={() => {
-                            selected_Sa_at = note as note_types;
-                            Sa_at_popup_status = false;
-                          }}>{note}</button
-                        >
-                      {/each}
-                    </div>
-                  {/snippet}
-                </Popover>
-              </div>
-              <PitchDisplay
-                audio_info={audio_info!}
-                bind:Sa_at={selected_Sa_at}
-                sargam_orientation={selected_sargam_orientation}
-                note_orientation={selected_note_orientation}
-              />
-              <!-- <PitchDisplay
-                  audio_info={{
-                    clarity: 100,
-                    note: 'D',
-                    scale: 0,
-                    detune: 0,
-                    pitch: 2
-                  }}
-                  bind:Sa_at={selected_Sa_at}
-                  sargam_orientation={selected_sargam_orientation}
-                  note_orientation={selected_note_orientation}
-                /> -->
-            </div>
-            <div class="text-3xl">{pitch} Hz</div>
-            <div class="space-y-0">
-              <div class="-mb-1 text-center text-sm font-semibold">Clarity</div>
-              <progress class="progress w-56" value={clarity} max="100"></progress>
-            </div>
-            <Popover
-              contentBase="card z-50 space-y-2 p-2 rounded-lg shadow-xl dark:bg-surface-900 bg-slate-100"
-              open={orientation_popup_status}
-              onOpenChange={(e) => (orientation_popup_status = e.open)}
-            >
-              {#snippet trigger()}
-                <div class="text-center font-bold outline-hidden">
-                  Orientation
-                  {#if !orientation_popup_status}
-                    <Icon src={BsChevronDown} class="text-lg" />
-                  {:else}
-                    <Icon src={BsChevronUp} class="text-lg" />
-                  {/if}
-                </div>
-              {/snippet}
-              {#snippet content()}
-                <div class="space-x-1">
-                  <span class="text-sm font-semibold">Sargam</span>
-                  <label>
-                    <input type="radio" bind:group={selected_sargam_orientation} value="vertical" />
-                    <span class="text-sm">Vertical</span>
-                  </label>
-                  <label>
-                    <input type="radio" bind:group={selected_sargam_orientation} value="radial" />
-                    <span class="text-sm">Radial</span>
-                  </label>
-                </div>
-                <div class="space-x-1">
-                  <span class="text-sm font-semibold">Note</span>
-                  <label>
-                    <input type="radio" bind:group={selected_note_orientation} value="vertical" />
-                    <span class="text-sm">Vertical</span>
-                  </label>
-                  <label>
-                    <input type="radio" bind:group={selected_note_orientation} value="radial" />
-                    <span class="text-sm">Radial</span>
-                  </label>
-                </div>
-              {/snippet}
-            </Popover>
-
-            <!-- Stop button -->
-            {@render stop_button()}
-          </div>
-        </div>
+        <CircularScale
+          audio_info={audio_info!}
+          bind:selected_Sa_at
+          bind:selected_sargam_orientation
+          bind:selected_note_orientation
+          {stop_button}
+        />
       </Tabs.Panel>
       <Tabs.Panel value="time_graph">
-        <div class="flex items-center justify-center">
-          <div
-            class="dow-md mr-4 rounded-lg bg-gradient-to-r from-amber-600 via-orange-700 to-red-600 px-3 py-1 text-base font-bold text-white select-none"
-          >
-            Alpha
-          </div>
-        </div>
-        <div class="mt-4 h-96 w-full">
-          {#if graphData.length > 1}
-            <div class="">
-              <!-- <h3 class="mb-4 text-lg font-semibold">Pitch Over Time</h3> -->
-              <svg class="h-80 w-full" viewBox="0 0 800 300">
-                <!-- Background -->
-                <rect width="800" height="300" fill="transparent" />
-
-                <!-- Grid lines for notes -->
-                {#each Array.from({ length: 13 }, (_, i) => i) as noteIndex}
-                  {@const y = (noteIndex / 12) * 240 + 30}
-                  <line
-                    x1="60"
-                    y1={y}
-                    x2="780"
-                    y2={y}
-                    stroke="#e5e7eb"
-                    stroke-width="1"
-                    opacity="0.5"
-                  />
-                  <text
-                    x="50"
-                    y={y + 4}
-                    text-anchor="end"
-                    class="fill-gray-600 text-xs dark:fill-gray-400"
-                  >
-                    {NOTE_ORDER_A_FIRST[12 - noteIndex - 1] || ''}
-                  </text>
-                {/each}
-
-                <!-- Time grid lines -->
-                {#each Array.from({ length: 6 }, (_, i) => i) as timeIndex}
-                  {@const x = (timeIndex / 5) * 720 + 60}
-                  <line
-                    x1={x}
-                    y1="30"
-                    x2={x}
-                    y2="270"
-                    stroke="#e5e7eb"
-                    stroke-width="1"
-                    opacity="0.3"
-                  />
-                  <!-- <text
-                    {x}
-                    y="290"
-                    text-anchor="middle"
-                    class="fill-gray-600 text-xs dark:fill-gray-400"
-                  >
-                    {(timeIndex * 0.5).toFixed(1)}s
-                  </text> -->
-                {/each}
-
-                <!-- Data line -->
-                {#if graphData.length > 1}
-                  {@const pathData = graphData
-                    .map((point, index) => {
-                      const x = (index / (MAX_PITCH_HISTORY_POINTS - 1)) * 720 + 60;
-                      const y = 270 - (point.y / 100) * 240;
-                      return index === 0 ? `M ${x} ${y}` : `L ${x} ${y}`;
-                    })
-                    .join(' ')}
-                  <path d={pathData} stroke="#3b82f6" stroke-width="2" fill="none" />
-
-                  <!-- Data points -->
-                  {#each graphData as point, index}
-                    {@const x = (index / (MAX_PITCH_HISTORY_POINTS - 1)) * 720 + 60}
-                    {@const y = 270 - (point.y / 100) * 240}
-                    <circle cx={x} cy={y} r="2" fill="#3b82f6" />
-                  {/each}
-
-                  <!-- Current frequency display -->
-                  {@const lastPoint = graphData[graphData.length - 1]}
-                  {@const lastX =
-                    ((graphData.length - 1) / (MAX_PITCH_HISTORY_POINTS - 1)) * 720 + 60}
-                  {@const lastY = 270 - (lastPoint.y / 100) * 240}
-                  <circle cx={lastX} cy={lastY} r="4" fill="#ef4444" />
-                  <text
-                    x={lastX + 10}
-                    y={lastY - 10}
-                    class="fill-gray-800 text-sm font-medium dark:fill-gray-200"
-                  >
-                    {lastPoint.pitch.toFixed(1)} Hz ({lastPoint.note}{audio_info?.scale})
-                  </text>
-                {/if}
-
-                <!-- Axes -->
-                <line x1="60" y1="30" x2="60" y2="270" stroke="#374151" stroke-width="2" />
-                <line x1="60" y1="270" x2="780" y2="270" stroke="#374151" stroke-width="2" />
-
-                <!-- Y-axis label -->
-                <text
-                  x="20"
-                  y="150"
-                  text-anchor="middle"
-                  transform="rotate(-90 20 150)"
-                  class="fill-gray-700 text-sm font-medium dark:fill-gray-300"
-                >
-                  Musical Notes
-                </text>
-
-                <!-- X-axis label -->
-                <!-- <text
-                  x="420"
-                  y="320"
-                  text-anchor="middle"
-                  class="fill-gray-700 text-sm font-medium dark:fill-gray-300"
-                >
-                  Time (seconds)
-                </text> -->
-              </svg>
-            </div>
-            {@render stop_button()}
-            <!-- {:else}
-            <div class="flex h-full items-center justify-center text-gray-500">
-              <div class="text-center">
-                <p class="text-lg font-medium">No data yet</p>
-                <p class="text-sm">Start recording to see the pitch graph</p>
-              </div>
-            </div> -->
-          {/if}
-        </div>
+        <PitchTimeGraph
+          {pitch_history}
+          {stop_button}
+          {MAX_PITCH_HISTORY_POINTS}
+          {AUDIO_INFO_UPDATE_INTERVAL}
+          audio_info_scale={audio_info?.scale}
+        />
       </Tabs.Panel>
     {/snippet}
   </Tabs>
